Organize files into util and style directories
