--- current-tangy-form-response-model.js
@@ -1,9 +1,0 @@
-class CurrentTangyFormResponseModel {
-  constructor(props) {
-    this._id = ''
-    this.collection = 'CurentTangyFormResponse'
-    this.formId = ''
-    this.responseId = ''
-    Object.assign(this, props)
-  }
-}
