+++ demo/index.html
@@ -15,22 +15,22 @@
       import '@polymer/iron-demo-helpers/demo-snippet';
     </script>
 
-    <script type="module" src="../tangy-form.js"></script>
-    <script type="module" src="../tangy-acasi.js"></script>
-    <script type="module" src="../tangy-box.js"></script>
-    <script type="module" src="../tangy-checkbox.js"></script>
-    <script type="module" src="../tangy-checkboxes.js"></script>
-    <script type="module" src="../tangy-complete-button.js"></script>
-    <script type="module" src="../tangy-eftouch.js"></script>
-    <script type="module" src="../tangy-gps.js"></script>
-    <script type="module" src="../tangy-input.js"></script>
-    <script type="module" src="../tangy-location.js"></script>
-    <script type="module" src="../tangy-overlay.js"></script>
-    <script type="module" src="../tangy-radio-button.js"></script>
-    <script type="module" src="../tangy-radio-buttons.js"></script>
-    <script type="module" src="../tangy-select.js"></script>
-    <script type="module" src="../tangy-timed.js"></script>
-    <script type="module" src="../tangy-toggle-button.js"></script>
+    <script type="module" src="../input/tangy-form.js"></script>
+    <script type="module" src="../input/tangy-acasi.js"></script>
+    <script type="module" src="../input/tangy-box.js"></script>
+    <script type="module" src="../input/tangy-checkbox.js"></script>
+    <script type="module" src="../input/tangy-checkboxes.js"></script>
+    <script type="module" src="../input/tangy-complete-button.js"></script>
+    <script type="module" src="../input/tangy-eftouch.js"></script>
+    <script type="module" src="../input/tangy-gps.js"></script>
+    <script type="module" src="../input/tangy-input.js"></script>
+    <script type="module" src="../input/tangy-location.js"></script>
+    <script type="module" src="../input/tangy-overlay.js"></script>
+    <script type="module" src="../input/tangy-radio-button.js"></script>
+    <script type="module" src="../input/tangy-radio-buttons.js"></script>
+    <script type="module" src="../input/tangy-select.js"></script>
+    <script type="module" src="../input/tangy-timed.js"></script>
+    <script type="module" src="../input/tangy-toggle-button.js"></script>
 
     <custom-style>
       <style is="custom-style" include="demo-pages-shared-styles">
