+++ input/tangy-box.js
@@ -1,9 +1,9 @@
 import { PolymerElement, html } from '@polymer/polymer/polymer-element.js';
 import { t } from './t.js'
-import './html-element-props.js'
+import '../util/html-element-props.js'
 import '@polymer/paper-radio-button/paper-radio-button.js'
-import './tangy-common-styles.js'
-import './tangy-element-styles.js'
+import '../style/tangy-common-styles.js'
+import '../style/tangy-element-styles.js'
 
     /**
      * `tangy-radio-button`
