+++ input/tangy-eftouch.js
@@ -1,9 +1,7 @@
 import { PolymerElement, html } from '@polymer/polymer/polymer-element.js';
-import { t } from './t.js'
-import './html-element-props.js'
+import '../util/html-element-props.js'
 import './tangy-radio-buttons.js'
-import './tangy-common-styles.js'
-import './tangy-eftouch-slide.js';
+import '../style/tangy-common-styles.js'
 
 /**
  * `tangy-acasi`
