+++ input/tangy-gps.js
@@ -1,8 +1,8 @@
 import { PolymerElement, html } from '@polymer/polymer/polymer-element.js';
-import { t } from './t.js'
-import './html-element-props.js'
-import './tangy-common-styles.js'
-import './tangy-element-styles.js'
+import { t } from '../util/t.js'
+import '../util/html-element-props.js'
+import '../style/tangy-common-styles.js'
+import '../style/tangy-element-styles.js'
 
 /**
  * `tangy-timed`
