+++ input/tangy-input.js
@@ -1,10 +1,10 @@
 import { PolymerElement, html } from '@polymer/polymer/polymer-element.js';
-import { t } from './t.js'
-import './html-element-props.js'
+import { t } from '../util/t.js'
+import '../util/html-element-props.js'
 import '@polymer/paper-input/paper-textarea.js'
 import '@polymer/paper-input/paper-input.js'
-import './tangy-common-styles.js'
-import './tangy-element-styles.js'
+import '../style/tangy-common-styles.js'
+import '../style/tangy-element-styles.js'
 
 /**
  * `tangy-input`
