+++ input/tangy-select.js
@@ -1,9 +1,8 @@
 import { PolymerElement, html } from '@polymer/polymer/polymer-element.js';
-import { t } from './t.js'
-import './html-element-props.js'
-import './tangy-element-styles.js';
-import './tangy-common-styles.js'
-import './mdc-select-style.js'
+import '../util/html-element-props.js'
+import '../style/tangy-element-styles.js';
+import '../style/tangy-common-styles.js'
+import '../style/mdc-select-style.js'
 /**
  * `tangy-select`
  *
