+++ style/mdc-select-style.js
