+++ tangy-complete-button.js
@@ -1,8 +1,7 @@
 import { PolymerElement, html } from '@polymer/polymer/polymer-element.js';
-import { t } from './t.js'
-import './html-element-props.js'
-import './tangy-element-styles.js';
-import './tangy-common-styles.js'
+import './util/html-element-props.js'
+import './style/tangy-element-styles.js';
+import './style/tangy-common-styles.js'
 /**
  * `tangy-complete-button`
  *
