+++ tangy-input-group.js
@@ -1,9 +1,8 @@
 import { PolymerElement, html } from '@polymer/polymer/polymer-element.js';
-import { t } from './t.js'
-import './html-element-props.js'
+import './util/html-element-props.js'
 import '@polymer/paper-card/paper-card.js'
-import './tangy-common-styles.js'
-import './tangy-element-styles.js'
+import './style/tangy-common-styles.js'
+import './style/tangy-element-styles.js'
 
     /**
      * `tangy-group`
