+++ tangy-input-groups.js
@@ -1,9 +1,8 @@
 import { PolymerElement, html } from '@polymer/polymer/polymer-element.js';
-import { t } from './t.js'
-import './html-element-props.js'
+import './util/html-element-props.js'
 import '@polymer/paper-checkbox/paper-checkbox.js'
-import './tangy-common-styles.js'
-import './tangy-element-styles.js'
+import './style/tangy-common-styles.js'
+import './style/tangy-element-styles.js'
 import './tangy-input-group.js'
 // https://stackoverflow.com/a/2117523/10139471
 function uuid() {
