+++ test/index.html
@@ -14,6 +14,7 @@
         'tangy-form_test.html',
         'tangy-form-item_test.html',
         'tangy-input-groups_test.html',
+        'tangy-radio-button_test.html',
         'tangy-radio-buttons_test.html',
         'tangy-location_test.html',
         'tangy-eftouch_test.html',
