+++ test/loc_test.html
@@ -13,7 +13,7 @@
   <body>
 
     <script type="module">
-      import { Loc } from '../loc.js'
+      import { Loc } from '../util/loc.js'
       import stub from './stub.js'
 
       suite('loc', () => {
