+++ test/tangy-eftouch_test.html
@@ -3,14 +3,10 @@
   <head>
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, minimum-scale=1, initial-scale=1, user-scalable=yes">
-
     <title>tangy-eftouch test</title>
-
     <script src="../node_modules/@webcomponents/webcomponentsjs/webcomponents-loader.js"></script>
     <script src="../node_modules/redux/dist/redux.js"></script>
     <script src="../node_modules/wct-browser-legacy/browser.js"></script>
-
-    <script type="module" src="../tangy-eftouch.js"></script>
   </head>
   <body>
 
@@ -116,6 +112,7 @@
 
     <script type="module">
       import * as Polymer from '../node_modules/@polymer/polymer/polymer-legacy.js'
+      import '../input/tangy-eftouch.js'
 
       suite('tangy-eftouch', () => {
