+++ test/tangy-form_test.html
@@ -175,8 +175,8 @@
       import * as Polymer from '../node_modules/@polymer/polymer/polymer-legacy.js'
       import stub from './stub.js' 
       import '../tangy-form.js'
-      import '../tangy-input.js'
-      import '../tangy-checkbox.js'
+      import '../input/tangy-input.js'
+      import '../input/tangy-checkbox.js'
       import { TangyFormResponseModel } from '../tangy-form-response-model.js'
 
       suite('TangyFormResponseModel', () => {
