+++ test/tangy-input-groups_test.html
@@ -3,13 +3,10 @@
   <head>
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, minimum-scale=1, initial-scale=1, user-scalable=yes">
-
     <title>tangy-input-groups test</title>
-
     <script src="../node_modules/@webcomponents/webcomponentsjs/webcomponents-loader.js"></script>
     <script src="../node_modules/redux/dist/redux.js"></script>
     <script src="../node_modules/wct-browser-legacy/browser.js"></script>
-
   </head>
   <body>
   
@@ -44,7 +41,7 @@
       import stub from './stub.js' 
       import '../tangy-form.js'
       import '../tangy-input-groups.js'
-      import '../tangy-input.js'
+      import '../input/tangy-input.js'
 
       suite('tangy-input-groups', () => {
         test('tangy-input-groups start with 3 groups', () => {
