+++ test/tangy-location_test.html
@@ -3,12 +3,9 @@
   <head>
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, minimum-scale=1, initial-scale=1, user-scalable=yes">
-
     <title>tangy-location test</title>
-
     <script src="../node_modules/@webcomponents/webcomponentsjs/webcomponents-loader.js"></script>
     <script src="../node_modules/wct-browser-legacy/browser.js"></script>
-
   </head>
   <body>
 
@@ -26,7 +23,7 @@
  
     <script type="module">
       import * as Polymer from '../node_modules/@polymer/polymer/polymer-legacy.js'
-      import '../tangy-location.js'
+      import '../input/tangy-location.js'
       suite('tangy-location', () => {
         test('should populate select list locations found in filterBy', (done) => {
           const element = fixture('TangyLocationFilterByFixture');
@@ -48,7 +45,6 @@
           }, {once: true})
         })
       })
-
     </script>
 
   </body>
