+++ test/tangy-radio-button_test.html
@@ -1,0 +1,28 @@
+<!doctype html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, minimum-scale=1, initial-scale=1, user-scalable=yes">
+    <title>tangy-radio-button test</title>
+    <script src="../node_modules/@webcomponents/webcomponentsjs/webcomponents-loader.js"></script>
+    <script src="../node_modules/redux/dist/redux.js"></script>
+    <script src="../node_modules/wct-browser-legacy/browser.js"></script>
+    <script type="module" src="../tangy-form.js"></script>
+  </head>
+  <body>
+    
+    <test-fixture id="TangyRadioButtonHideButtonFixture">
+      <template>
+        <tangy-radio-button name="foo" hide-button>Foo</tangy-radio-button>
+      </template>
+    </test-fixture>
+
+    <script type="module">
+      import '../input/tangy-radio-button.js'
+      suite('tangy-radio-button', () => {
+        test('should hide button')
+      })
+    </script>
+
+  </body>
+</html>
